requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
requests==2.26.0
